День 6

+++ AI Challenge 5/Models/RequestModel.swift
@@ -8,7 +8,7 @@
 nonisolated struct RequestModel: Encodable, Sendable {
     let model: GigaChatModel
     let messages: [Message]
-    let temperature: Int
+    let temperature: Float
     let maxTokens: Int
     let repetitionPenalty: Float
     let updateInterval: Int
